bug discovered, private key QR loaded the public key
_to EOS cafe: verify this before you merge this pull request_

**How are users affected?**
Paper wallets that have been generated so far with https://eoscafe.github.io/eos-paper-wallet/ are still valid, except that the QR for the private key is not the private key, it is a duplicate of the public key. Paper wallets generated after this pull request is merged will have both a public key and a private key QR code.